Delegate construction to `DefaultEntityFactory`
